tried one way;seems good! store just to rename repos
